Move error classes into AwsMfa::Errors
