+++ lib/aws_mfa.rb
@@ -1,13 +1,8 @@
 require 'json'
+require 'aws_mfa/errors'
 
 class AwsMfa
   attr_reader :aws_config_dir
-
-  class Error < StandardError; end
-  class ConfigurationNotFound < Error; end
-  class CommandNotFound < Error; end
-  class DeviceNotFound < Error; end
-  class InvalidCode < Error; end
 
   def initialize
     validate_aws_installed
@@ -15,7 +10,7 @@ class AwsMfa
   end
 
   def validate_aws_installed
-    raise CommandNotFound, 'Could not find the aws command' unless which('aws')
+    raise Errors::CommandNotFound, 'Could not find the aws command' unless which('aws')
   end
 
   def find_aws_config
@@ -28,7 +23,7 @@ class AwsMfa
     end
 
     unless File.readable?(aws_config_file)
-      raise ConfigurationNotFound, 'Aws configuration not found. You must run `aws cli configure`'
+      raise Errors::ConfigurationNotFound, 'Aws configuration not found. You must run `aws cli configure`'
     end
 
     aws_config_dir
@@ -68,7 +63,7 @@ class AwsMfa
     if mfa_devices.any?
       mfa_devices.first.fetch('SerialNumber')
     else
-      raise DeviceNotFound, 'No MFA devices were found for your account'
+      raise Errors::DeviceNotFound, 'No MFA devices were found for your account'
     end
   end
 
@@ -111,7 +106,7 @@ class AwsMfa
   def request_code_from_user
     puts 'Enter the 6-digit code from your MFA device:'
     code = $stdin.gets.chomp
-    raise InvalidCode, 'That is an invalid MFA code' unless code =~ /^\d{6}$/
+    raise Errors::InvalidCode, 'That is an invalid MFA code' unless code =~ /^\d{6}$/
     code
   end
