+++ lib/aws_mfa/errors.rb
@@ -1,0 +1,11 @@
+class AwsMfa
+  class Errors
+
+    class Error < StandardError; end
+    class ConfigurationNotFound < Error; end
+    class CommandNotFound < Error; end
+    class DeviceNotFound < Error; end
+    class InvalidCode < Error; end
+
+  end
+end
